Get things working again with new protocol changes
@aslakhellesoy the change here on line 26 of `renderer/index.js` is an
example of what I'm talking about with the locations being more
convenient to use with a `uri`. It's a chore to have to check
whether the location has a `uri` or not, and if not fall back to the
`uri` of the test case.

Ideally really I think I'd like to be able to load this JSON into a
value object where I could just call `toString()` on the location. Do
you think that's something we could build into the Gherkin library and
ship with it?